requirements.txt
Django==5.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
gunicorn==23.0.0